Add files via upload
The file Salaries of MLB Players.xlsx lists the salaries of MLB players in the years 2007 to 2009. Each row corresponds to 
a particular player. As indicated by blank salaries, some players played in one of these years, some 
played in two of these years, and the rest played in all three years